+++ sdk/identity/identity/src/credentials/defaultAzureCredentialFunctions.ts
@@ -29,7 +29,7 @@ export function createDefaultManagedIdentityCredential(
   options:
     | DefaultAzureCredentialOptions
     | DefaultAzureCredentialResourceIdOptions
-    | DefaultAzureCredentialClientIdOptions = {},
+    | DefaultAzureCredentialClientIdOptions = {}
 ): TokenCredential {
   options.retryOptions ??= {
     maxRetries: 5,
@@ -61,7 +61,7 @@ export function createDefaultManagedIdentityCredential(
 
     return new ManagedIdentityCredential(
       workloadIdentityClientId,
-      workloadIdentityCredentialOptions,
+      workloadIdentityCredentialOptions
     );
   }
 
@@ -85,7 +85,7 @@ export function createDefaultManagedIdentityCredential(
  * @internal
  */
 export function createDefaultWorkloadIdentityCredential(
-  options?: DefaultAzureCredentialOptions | DefaultAzureCredentialClientIdOptions,
+  options?: DefaultAzureCredentialOptions | DefaultAzureCredentialClientIdOptions
 ): TokenCredential {
   const managedIdentityClientId =
     (options as DefaultAzureCredentialClientIdOptions)?.managedIdentityClientId ??
@@ -123,7 +123,7 @@ export function createDefaultWorkloadIdentityCredential(
  * @internal
  */
 export function createDefaultAzureDeveloperCliCredential(
-  options: DefaultAzureCredentialOptions = {},
+  options: DefaultAzureCredentialOptions = {}
 ): TokenCredential {
   return new AzureDeveloperCliCredential(options);
 }
@@ -135,9 +135,9 @@ export function createDefaultAzureDeveloperCliCredential(
  * @internal
  */
 export function createDefaultAzureCliCredential(
-  options: DefaultAzureCredentialOptions = {},
+  options: DefaultAzureCredentialOptions = {}
 ): TokenCredential {
-  return new AzureCliCredential({ ...options });
+  return new AzureCliCredential(options);
 }
 
 /**
@@ -147,9 +147,9 @@ export function createDefaultAzureCliCredential(
  * @internal
  */
 export function createDefaultAzurePowershellCredential(
-  options: DefaultAzureCredentialOptions = {},
+  options: DefaultAzureCredentialOptions = {}
 ): TokenCredential {
-  return new AzurePowerShellCredential({ ...options });
+  return new AzurePowerShellCredential(options);
 }
 
 /**
@@ -159,7 +159,7 @@ export function createDefaultAzurePowershellCredential(
  * @internal
  */
 export function createEnvironmentCredential(
-  options: DefaultAzureCredentialOptions = {},
+  options: DefaultAzureCredentialOptions = {}
 ): TokenCredential {
   return new EnvironmentCredential(options);
 }